Started work on making a britecharts API so we can produce charts alot quicker
